fixed linux build warnings
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3392 406a0c4d-033a-0410-8de8-e80135713968

+++ src/searchd.cpp
@@ -14834,7 +14834,7 @@ static void UservarsThreadFunc ( void * )
 
 	char dBuf[512];
 	const int iMaxString = 80;
-	assert ( sizeof ( dBuf )>iMaxString );
+	assert ( (int)sizeof(dBuf) > iMaxString );
 
 	CSphString sError;
 	CSphWriter tWriter;
